+++ packages/web3-eth-accounts/src/account.ts
@@ -1,6 +1,6 @@
 import { utils, getPublicKey } from 'ethereum-cryptography/secp256k1';
-import { toChecksumAddress, bytesToHex, sha3Raw } from 'web3-utils';
-import { PrivateKeyError } from './errors';
+import { toChecksumAddress, bytesToHex, sha3Raw, HexString } from 'web3-utils';
+import { InvalidPrivateKeyError, PrivateKeyLengthError } from './errors';
 // Will be added later
 export const encrypt = (): boolean => true;
 
@@ -14,7 +14,7 @@ export const signTransaction = (): boolean => true;
  * Get address from private key
  */
 export const privateKeyToAccount = (
-	privateKey: string | Uint8Array,
+	privateKey: string | Buffer,
 ): {
 	address: string;
 	privateKey: string;
@@ -22,16 +22,22 @@ export const privateKeyToAccount = (
 	sign: () => boolean;
 	encrypt: () => boolean;
 } => {
+	if (!privateKey) {
+		throw new InvalidPrivateKeyError(privateKey);
+	}
+
 	const stringPrivateKey =
 		typeof privateKey === 'object' ? Buffer.from(privateKey).toString('hex') : privateKey;
+
 	const updatedKey = stringPrivateKey.startsWith('0x')
 		? stringPrivateKey.slice(2)
 		: stringPrivateKey;
 
 	// Must be 64 hex characters
 	if (updatedKey.length !== 64) {
-		throw new PrivateKeyError(updatedKey);
+		throw new PrivateKeyLengthError(updatedKey);
 	}
+
 	const publicKey = getPublicKey(updatedKey);
 
 	const publicKeyString = `0x${publicKey.slice(2)}`;
@@ -45,11 +51,11 @@ export const privateKeyToAccount = (
  * Returns a random hex string by the given bytes size
  */
 export const create = (): {
-	address: string;
+	address: HexString;
 	privateKey: string;
-	signTransaction: Function; // From 1.x
-	sign: Function;
-	encrypt: Function;
+	signTransaction: () => boolean; // From 1.x
+	sign: () => boolean;
+	encrypt: () => boolean;
 } => {
 	const privateKey = utils.randomPrivateKey();
 	const address = getPublicKey(privateKey);