+++ CHANGELOG.md
@@ -340,6 +340,7 @@ Released with 1.0.0-beta.37 code base.
 1. The following functions `soliditySha3` `soliditySha3Raw` `encodePacked` now includes type validation and requires type specficiation, instead of guessing the value type
 2. The functions `soliditySha3` `soliditySha3Raw` `encodePacked` does not support BN and now supports `BigInt`
 3. The functions `flattenTypes` and `jsonInterfaceMethodToString` moved to the `web3-eth-abi` package
+4. The function `isAddress` now includes an optional parameter `checkChecksum` type boolean
 
 ### web3-eth-accounts