+++ CHANGELOG.md
@@ -478,8 +478,6 @@ Released with 1.0.0-beta.37 code base.
 - Correct `web3.rst` example in documentation (#4511)
 - Correct `BlockHeader` typing (`receiptRoot` -> `receiptsRoot`) (#4452)
 
-## [Unreleased]
-
 ## [1.7.0]
 
 ### Added
@@ -492,3 +490,7 @@ Released with 1.0.0-beta.37 code base.
 ### Changed
 
  - Changed getFeeHistory first parameter type from `number` to `hex` according to the [spec](https://playground.open-rpc.org/?schemaUrl=https://raw.githubusercontent.com/ethereum/eth1.0-apis/assembled-spec/openrpc.json&uiSchema%5BappBar%5D%5Bui:splitView%5D=false&uiSchema%5BappBar%5D%5Bui:input%5D=false&uiSchema%5BappBar%5D%5Bui:examplesDropdown%5D=false) (#4529)
+
+## [Unreleased]
+
+## [1.8.0]